\name{ALLbetas}
\docType{data}
\alias{ALLbetas}
\title{EPICv2 beta values for DMR calling}
\description{
  Matrix of EPICv2 beta values from Noguera-Castells et al. (2023) consisting of five B cell acute lymphoblastic leukaemia (BALL) and five T cell acute lymphoblastic leukaemia (TALL) samples for DMR calling.
}
\usage{data(ALLbetas)}
\format{data.frame}
\source{
  \url{https://ftp.ncbi.nlm.nih.gov/geo/series/GSE222nnn/GSE222919/suppl/GSE222919_processed_data.txt.gz}
}
\keyword{datasets}