\name{crosshyb}
\docType{data}
\alias{crosshyb}
\title{Potentially cross-hybridising Illumina probes}
\description{
  This is a character vector of Illumina probes whose probe sequence promiscuously aligns to non-target sections of the genome with a matching of 47bp or higher.
}
\usage{data(crosshyb)}
\format{character}
\source{
  \url{https://static-content.springer.com/esm/art\%3A10.1186\%2Fs13059-016-1066-1/MediaObjects/13059_2016_1066_MOESM2_ESM.csv} (accessed September 2019)   
  \url{https://static-content.springer.com/esm/art\%3A10.1186\%2Fs13059-016-1066-1/MediaObjects/13059_2016_1066_MOESM3_ESM.csv} (accessed September 2019)
  \url{http://www.sickkids.ca/MS-Office-Files/Research/Weksberg Lab/48639-non-specific-probes-Illumina450k.xlsx} (accessed September 2019)}
\keyword{datasets}