\name{DMRcatedata-package}
\alias{DMRcatedata-package}
\alias{DMRcatedata}
\docType{package}
\title{
Dataset to use with the DMRcate Pipeline 
}
\description{
SNP annotation, cross-hybridising probes and sample 450k data
}
\author{
Tim Peters, CSIRO Computational Informatics 

Maintainer: Tim Peters<Tim.Peters@csiro.au>
}

\examples{
data(dmrcatedata)
}

\keyword{datasets}
