\name{dmrcatedata}
\alias{dmrcatedata}
\docType{data}
\title{Accompanying data package for DMRcate}
\description{
SNP annotation, cross-hybridising probes and sample 450k data
}
\usage{data(dmrcatedata)}
\format{matrix (\code{myBetas}), factor (\code{crosshyb}) and data.frame (\code{illuminaSNPs})}
\source{\code{myBetas} sourced from The Cancer Genome Atlas (TCGA) data repository, colon and rectal adenocarcinoma; \code{illuminaSNPs} sourced from \url{http://supportres.illumina.com/documents/myillumina/88bab663-307c-444a-848e-0ed6c338ee4d/humanmethylation450_15017482_v.1.2.snpupdate.table.v3.txt}, accessed February 2014; \code{crosshyb} sourced from \url{http://www.sickkids.ca/MS-Office-Files/Research/Weksberg Lab/48639-non-specific-probes-Illumina450k.xlsx}, accessed February 2014}
\keyword{datasets}