\name{epicv2snps}
\docType{data}
\alias{epicv2snps}
\title{SNP information for EPICv2 Illumina probes}
\description{
  This is a data.frame of Illumina probes from EPICv2 whose target CpG lies on or near a known SNP. SNP ID, distance and minor allele frequency are all reported here.
}
\usage{data(snpsall)}
\format{data.frame}
\source{
  \url{https://static-content.springer.com/esm/art\%3A10.1186\%2Fs12864-024-10027-5/MediaObjects/12864_2024_10027_MOESM4_ESM.csv}
}
\keyword{datasets}