\name{snpsall}
\docType{data}
\alias{snpsall}
\title{SNP information for EPICv1 and 450K Illumina probes}
\description{
  This is a data.frame of Illumina probes from EPICv1 and 450K whose target CpG lies on or near a known SNP. SNP ID, distance and minor allele frequency are all reported here.
}
\usage{data(snpsall)}
\format{data.frame}
\source{
  \url{https://static-content.springer.com/esm/art\%3A10.1186\%2Fs13059-016-1066-1/MediaObjects/13059_2016_1066_MOESM4_ESM.csv}
       (accessed September 2019)   
  \url{https://static-content.springer.com/esm/art\%3A10.1186\%2Fs13059-016-1066-1/MediaObjects/13059_2016_1066_MOESM5_ESM.csv} (accessed September 2019)
  \url{https://static-content.springer.com/esm/art\%3A10.1186\%2Fs13059-016-1066-1/MediaObjects/13059_2016_1066_MOESM6_ESM.csv} (accessed September 2019)
  \url{http://www.sickkids.ca/MS-Office-Files/Research/Weksberg Lab/48640-polymorphic-CpGs-Illumina450k.xlsx} (accessed September 2019)}
\keyword{datasets}