\name{DMRcatedata-package}
\alias{DMRcatedata-package}
\alias{DMRcatedata}
\docType{package}
\title{
Dataset to use with the DMRcate pipeline 
}
\description{
SNP annotation, cross-hybridising probes, XY probes, and gene annotation for hg19, hg38 and mm10
}
\author{
Tim Peters, Immunogenomics Laboratory, Garvan Institute of Medical Research

Maintainer: Tim Peters<t.peters@garvan.org.au>
}

\examples{
data(dmrcatedata)
}

\keyword{datasets}
