\name{DMRcatedata-package}
\alias{DMRcatedata-package}
\alias{DMRcatedata}
\docType{package}
\title{
Dataset to use with the DMRcate Pipeline 
}
\description{
SNP annotation, cross-hybridising probes and sample 450k data
}
\author{
Tim Peters, Epigenetics Program, Garvan Institute of Medical Research

Maintainer: Tim Peters<t.peters@garvan.org.au>
}

\examples{
data(dmrcatedata)
}

\keyword{datasets}
