\name{epicv2snps}
\docType{data}
\alias{epicv2snps}
\title{SNP information for EPICv2 Illumina probes}
\description{
  This is a data.frame of Illumina probes from EPICv2 whose target CpG lies on or near a known SNP. SNP ID, distance and minor allele frequency are all reported here.
}
\usage{data(snpsall)}
\format{data.frame}
\source{
  \url{https://bioconductor.org/packages/release/data/experiment/html/DMRcatedata.html}
}
\keyword{datasets}