\name{hg38.generanges}
\docType{data}
\alias{hg38.generanges}
\title{Start and stop positions of all genes in hg38}
\description{
  This data set gives the genomic intervals of all gene regions in the Ensembl Release 96 of hg38.
}
\usage{data(hg38.generanges)}
\format{A GRanges object with 57316 intervals.}
\source{\url{ftp://ftp.ensembl.org/pub/release-96/gtf/homo_sapiens/Homo_sapiens.GRCh38.96.chr.gtf.gz}}
\keyword{datasets}