\name{mm10.generanges}
\docType{data}
\alias{mm10.generanges}
\title{Start and stop positions of all genes in mm10}
\description{
  This data set gives the genomic intervals of all gene regions in the Ensembl Release 96 of mm10.
}
\usage{data(dmrcatedata)}
\format{A GRanges object with 55341 intervals.}
\source{\url{ftp://ftp.ensembl.org/pub/release-96/gtf/mus_musculus/Mus_musculus.GRCm38.96.chr.gtf.gz}}
\keyword{datasets}